Daily commit after work on CJFAS short comm and Fish and Fisheries manuscript.
I reviewed the makefile for the Fish and Fisheries manuscript and also worked on the dynamics generation of the assessment references in Table S1. After this commit I will start the long and painful process of going through each assessment in the database, fix its RIS entry so that the resulting citation appears correctly in Table S1 of the F&F manuscript. I will also try an close as many assessments as possible through this process and finalise the QAQC process.