Commit before starting the review work for the F&F manuscript.
Olaf is coming to Dal next week and we will likely do some pretty active database work, so I'm getting this commit to capture the snapshot prior to all the changes. I will likely take another commit later this week and include a fresh database dump as well. I hope to migrate the database to a production machine prior to the publication of the manuscript so I might also perform database upgrade to 8.4 before migration.