Commit after sending the fir revised draft of the paper to co-authors.
I have done some modification related to the F&F manuscript and this commit aims to capture that. I found out that COD4VsW wasn't making it into the mansucript because it had assess=0. Also, found out that ANDERSON Atlantic surfclam was a duplicate, so removed that, and the current stock's catch data was out by 1E03!