Commit to capture the code I wrote to generate the "final" version of the results and figures for the Fish and Fisheries paper.
I created new directories called "PROOFS-final" for both R and tex in the folder for the Fish and Fisheries paper. I want to make sure that I can rerun this code using the most up-to-date version of the database, while also maintaining an archived version of the results and figures.