End-of-day commit after sending the corrected proofs to FAF.
I removed the exploitation reference point for sablefish Canadian Pacific Coast, whic had repercussions on the whole FAF paper. I also fiddled around with updating the Figure 1 map. This is about as close as it will get to the production version so I will take a database backup and restore it on nautilus-vm shortly.